backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
PyJWT
email-validator
flasgger
flask-migrate
pytest
gunicorn
flask-cors==6.0.1
prometheus_flask_exporter
Flask-Limiter
Flask-Limiter[redis]
redis
sentry-sdk
python-json-logger
flask-caching
pytest-mock
langchain
chromadb
openai        
sentence-transformers
transformers 